requirements.txt
flask==2.3.3
pyyaml==6.0.1
pytest==7.4.0
pytest-cov==4.1.0
requests==2.31.0
gunicorn==21.2.0